refactor utils and code clean: * move files utils into one file * adjust imports * fix usage * add todo comment.
