+++ components/ui/Spinner.tsx
@@ -1,6 +1,7 @@
-import formatFileSize from "@/lib/files/formatFileSize";
-import { Progress } from "@/components/ui/progress";
 import { useEffect, useState } from "react";
+
+import { Progress } from "@/components/ui/progress";
+import { formatFileSize } from "@/lib/files";
 
 interface SpinnerProps {
   total?: number;
