--- lib/files/calculateTotalFilesSize.ts
@@ -1,2 +1,0 @@
-export default (files: File[]) =>
-  files.reduce((total, file) => total + file.size, 0);
